fix: `getData` function should only change when `getOptions` changes
Updates the `useVisitorData` hook to only return a new `getData`
function when the `getOptions` parameter is semantically different.
Previously, the `getData` function would be recreated if the
`getOptions` parameter was a different object instance but otherwise
equivalent to the last invocation.

The fix stores the `getOptions` parameter as a property of the state to
ensure that the value used as a dependency in the `getCallback` call
only changes when the semantics of the options for the `get` call
change. The state is updated using a pattern to update the state of
a component during a render by guarding that state update with a
conditional. More information about this pattern can be found in the
React docs:

https://react.dev/reference/react/useState#storing-information-from-previous-renders

The `usePrevious` utility was no longer used so it was removed.

A new unit test is also added that validates the expected behavior of
the `getData` function returned by `useVisitorData`.

Fixes #132
Fixes INTER-1277